[FI-561] price Long타입으로 수정, 회차 등록시 회차 좌석 자동 생성 구현

+++ FLOWIN/performance-service/src/main/java/io/why503/performanceservice/domain/roundSeat/model/dto/response/SeatReserveResponse.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public record SeatReserveResponse(
         Long roundSeatSq,       // 티켓과 연결될 ID
         String roundSeatStatus, // 상태
-        Integer price,          // 가격
+        Long price,          // 가격
         String grade,           // 등급
         String seatArea,        // 구역
         Integer seatAreaNumber,  // 번호
