+++ FLOWIN/performance-service/src/main/java/io/why503/performanceservice/domain/showseat/model/dto/response/ShowSeatResponse.java
@@ -7,6 +7,6 @@ public record ShowSeatResponse(
         Long showSeatSq,        // ShowSeat PK
         Long showSq,            // 공연 PK (순환참조 방지용)
         ShowSeatGrade grade,    // 등급
-        int price,              // 가격
+        Long price,              // 가격
         SeatResponse seat       // 물리적 좌석 정보
 ){}
